windows compile fix.

+++ src/maidsafe/vault/sync.h
@@ -110,7 +110,7 @@ bool IsResolved(const UnresolvedAction& unresolved_action) {
   // Received syncs from at least majority of peers(including itself)
   // Shall have tri-state : unresolved, resolved, already-resolved
   // However, as IsResolved get called only once, here use strict "==" to differentiate
-  uint16_t total_recieved = static_cast<uint32_t>(unresolved_action.peer_and_entry_ids.size() +
+  uint16_t total_recieved = static_cast<uint16_t>(unresolved_action.peer_and_entry_ids.size() +
            (unresolved_action.this_node_and_entry_id ? 1U : 0U));
   LOG(kVerbose) << "IsResolved  total_recieved : " << total_recieved;
   return (total_recieved == ((routing::Parameters::node_group_size / 2) + 1U));
